requirements.txt
django==1.8.2
django-storages==1.1.8
django-compressor==1.5
boto==2.38.0